+++ src/labkey/Experiment.ts
@@ -597,40 +597,63 @@ export function saveRuns(options: SaveRunsOptions): XMLHttpRequest {
     });
 }
 
-export interface GenIdActionsOptions extends RequestCallbackOptions {
+export interface EntitySequenceActionsOptions extends RequestCallbackOptions {
     containerPath?: string;
-    genId?: number;
-    kindName: 'SampleSet' | 'DataClass';
-    rowId: number;
+    seqType: 'genId' | 'rootSampleCount' | 'sampleCount';
+    newValue?: number;
+    kindName?: 'SampleSet' | 'DataClass';
+    rowId?: number;
 }
+
 /**
- * Set the current genId sequence value for the data type (sampleset or dataclass)
+ * Update the current value for the DB Sequences used in sample or data name expression
+ *
+ * #### Examples
+ *
+ * ```js
+ * LABKEY.Experiment.setEntitySequence({
+ *     seqType: 'genId',
+ *     newValue: 100,
+ *     kindName: 'SampleSet',
+ *     rowId: 12
+ * });
+ *
+ * LABKEY.Experiment.setEntitySequence({
+ *     seqType: 'sampleCount',
+ *     newValue: 100,
+ *     containerPath: 'SamplesFolder/'
+ * });
+ *
+ * ```
+ *
  */
-export function setGenId(config: GenIdActionsOptions): XMLHttpRequest {
+export function setEntitySequence(config: EntitySequenceActionsOptions): XMLHttpRequest {
     return request({
-        url: buildURL('experiment', 'setGenId.api', config.containerPath),
+        url: buildURL('experiment', 'setEntitySequence.api', config.containerPath),
         method: 'POST',
         success: getCallbackWrapper(getOnSuccess(config), config.scope),
         failure: getCallbackWrapper(getOnFailure(config), config.scope, true),
         jsonData: {
             rowId: config.rowId,
             kindName: config.kindName,
-            genId: config.genId,
+            newValue: config.newValue,
+            seqType: config.seqType,
         },
     });
 }
 
 /**
- * Get the current genId for the data type (sampleset or dataclass)
+ * Get the current value for the DB Sequences used in sample or data name expression
  */
-export function getGenId(config: GenIdActionsOptions): XMLHttpRequest {
+export function getEntitySequence(config: EntitySequenceActionsOptions): XMLHttpRequest {
     return request({
-        url: buildURL('experiment', 'getGenId.api', config.containerPath),
+        url: buildURL('experiment', 'getEntitySequence.api', config.containerPath),
         success: getCallbackWrapper(getOnSuccess(config), config.scope),
         failure: getCallbackWrapper(getOnFailure(config), config.scope, true),
         params: {
             rowId: config.rowId,
             kindName: config.kindName,
+            seqType: config.seqType,
         },
     });
 }